+++ tests/test_checker/test_noqa.py
@@ -91,6 +91,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS336': 1,
     'WPS337': 1,
     'WPS338': 1,
+    'WPS339': 1,
 
     'WPS400': 0,
     'WPS401': 0,