+++ tests/test_visitors/test_tokenize/test_primitives/test_numbers/test_number_meaningless_zeros.py
@@ -3,7 +3,7 @@
 import pytest
 
 from wemake_python_styleguide.violations.consistency import (
-    BadNumberSuffixViolation,
+    NumberWithMeaninglessZeroViolation,
 )
 from wemake_python_styleguide.visitors.tokenize.primitives import (
     WrongNumberTokenVisitor,
@@ -11,31 +11,51 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
 
 
 @pytest.mark.parametrize('number', [
-    '0X1',
-    '0X1A',
-    '0XFF',
-    '0Xe',
-    '1.5E+10',
-    '0O11',
-    '0B1001',
+    '0x01',
+    '0x00A',
 
-    '-0X1',
-    '-0X1A',
-    '-0XFF',
-    '-0Xe',
-    '-1.5E+10',
-    '-0O11',
-    '-0B1001',
+    '0e00',
+    '0e01',
+    '1.5e+010',
+    '1.5e-010',
 
-    '+0X1',
-    '+0X1A',
-    '+0XFF',
-    '+0Xe',
-    '+1.5E+10',
-    '+0O11',
-    '+0B1001',
+    '0o01',
+    '0o00007',
+
+    '0b0001',
+    '0b01',
+
+    '-0x01',
+    '-0x00A',
+
+    '-0e00',
+    '-0e01',
+    '-1.5e-010',
+
+    '-0o01',
+    '-0o00007',
+
+    '-0b0001',
+    '-0b01',
+
+    '+0x01',
+    '+0x00A',
+
+    '+0e00',
+    '+0e01',
+    '+1.5e+010',
+    '+1.5e-010',
+
+    '-0o01',
+    '-0o00007',
+
+    '-0b0001',
+    '-0b01',
+    '--0b0001',
+    '++0b01',
+    '-+0b0001',
 ])
-def test_bad_number_suffixes(
+def test_meaningless_zeros(
     parse_tokens,
     assert_errors,
     assert_error_text,
@@ -50,41 +70,63 @@ def test_bad_number_suffixes(
     visitor = WrongNumberTokenVisitor(default_options, file_tokens=file_tokens)
     visitor.run()
 
-    assert_errors(visitor, [BadNumberSuffixViolation])
+    assert_errors(visitor, [NumberWithMeaninglessZeroViolation])
     assert_error_text(visitor, number.lstrip('-').lstrip('+'))
 
 
 @pytest.mark.parametrize('number', [
-    '0xFF',
+    '0x0',
+    '0x10',
+    '0xA00',
+
+    '0e0',
+    '0e10',
     '1.5e+10',
-    '0o11',
-    '0b1001',
+    '1.5e-100',
 
-    '-0xAF',
-    '-3e+10',
-    '-0o11',
-    '-0b1111',
+    '0o0',
+    '0o1',
+    '0o7000',
 
-    '+0xFF',
+    '0b0',
+    '0b1',
+    '0b100000',
+
+    '-0x0',
+    '-0x10',
+    '-0xA00',
+
+    '-0e0',
+    '-0e10',
+    '-1.5e+10',
+    '-1.5e-100',
+
+    '-0o0',
+    '-0o1',
+    '0o7000',
+
+    '-0b0',
+    '-0b1',
+    '-0b100000',
+
+    '+0x0',
+    '+0x10',
+    '+0xA00',
+
+    '+0e0',
+    '+0e10',
     '+1.5e+10',
-    '+0o11',
-    '+0b1001',
+    '+1.5e-100',
 
-    # Regression for 557:
-    # https://github.com/wemake-services/wemake-python-styleguide/issues/557
-    '0xE',
-    '0xe',
-    '-0xE',
-    '-0xe',
-    '-0xB',
-    '+0xE',
-    '+0xe',
-    '+0xEE',
-    '0xEEE',
-    '0x1E',
-    '0xE1',
+    '+0o0',
+    '+0o1',
+    '+0o7000',
+
+    '+0b0',
+    '+0b1',
+    '+0b100000',
 ])
-def test_correct_number_suffixes(
+def test_correct_zeros(
     parse_tokens,
     assert_errors,
     default_options,
@@ -102,25 +144,23 @@ def test_correct_number_suffixes(
 
 
 @pytest.mark.parametrize('code', [
-    'print({0})',
-    'regular = {0}',
     'as_string = "{0}"',
 ])
 @pytest.mark.parametrize('number', [
-    '0xFF',
-    '1.5e+10',
-    '0o11',
-    '0b1001',
+    '0x0FF',
+    '1.5e+01',
+    '0o011',
+    '0b01001',
 
-    '-0xAF',
-    '-3e+10',
-    '-0o11',
-    '-0b1111',
+    '-0x0FF',
+    '-1.5e+01',
+    '-0o011',
+    '-0b01001',
 
-    '+0xAF',
-    '+1.3e+10',
-    '+0o11',
-    '+0b1111',
+    '+0x0FF',
+    '+1.5e+01',
+    '+0o011',
+    '+0b01001',
 ])
 def test_similar_strings(
     parse_tokens,